--- .travis.yml
@@ -1,10 +1,0 @@
-language: node_js
-node_js:
-  - '8'
-  - '10'
-
-env:
-  - NODE_ENV=test
-script:
-  - npm run test:ci
-  - npm run lint
