+++ .vscode/settings.json
@@ -1,6 +1,5 @@
 {
   "editor.formatOnSave": true,
-
   // turn VSCode internal formatter off for JS and JSX, ESLint will do it.
   "[javascript]": {
     "editor.formatOnSave": false
@@ -8,10 +7,8 @@
   "[javascriptreact]": {
     "editor.formatOnSave": false
   },
-
   "eslint.validate": ["javascript", "javascriptreact", "html"],
   "eslint.autoFixOnSave": true,
-
   // Optional BUT IMPORTANT: If you have the prettier extension enabled for other languages like CSS and HTML, turn it off for JS since we are doing it through Eslint already
   "prettier.disableLanguages": ["javascript", "javascriptreact"]
 }
