+++ lerna.json
@@ -1,6 +1,6 @@
 {
-  "packages": [
-    "web"
-  ],
-  "version": "0.0.1"
+  "npmClient": "yarn",
+  "useWorkspaces": true,
+  "packages": ["packages/*"],
+  "version": "0.0.0"
 }
