+++ now.json
@@ -2,21 +2,23 @@
   "version": 2,
   "name": "adify",
   "public": true,
-  "alias": [
-    "adify"
-  ],
+  "alias": ["adify"],
   "scope": "iqubex-technologies",
   "build": {
     "env": {
-
+      "TEST": "TEST"
     }
   },
-  "builds": [{
-    "src": "web/package.json",
-    "use": "@now/next"
-  }],
-  "routes": [{
-    "src": "/(.*)",
-    "dest": "/web/$1"
-  }]
+  "builds": [
+    {
+      "src": "packages/web/next.config.js",
+      "use": "@now/next"
+    }
+  ],
+  "routes": [
+    {
+      "src": "/(.*)",
+      "dest": "/web/$1"
+    }
+  ]
 }
