+++ packages/web/.env.example
@@ -1,0 +1,2 @@
+# Rename me to .env
+TEST= it works from .env
