+++ packages/web/.eslintignore
@@ -1,0 +1,1 @@
+# add files to ignore here
