+++ packages/web/.eslintrc.js
@@ -1,4 +1,4 @@
-{
+module.exports = {
   "extends": [
     "wesbos"
   ],
