+++ packages/web/next.config.js
@@ -1,0 +1,8 @@
+const withSass = require('@zeit/next-sass')
+
+module.exports = withSass({
+  target: 'serverless',
+  env: {
+    TEST: process.env.TEST,
+  },
+})
