+++ packages/web/pages/post/[pid].js
@@ -1,0 +1,12 @@
+// dynamic routing example which eliminates the need of an express server
+import React from 'react'
+import { useRouter } from 'next/router'
+
+const Post = () => {
+  const router = useRouter()
+  const { pid } = router.query
+
+  return <p>Post: {pid}</p>
+}
+
+export default Post
