--- web/package.json
@@ -1,59 +1,0 @@
-{
-  "name": "@adify-monorepo/web",
-  "version": "0.0.1",
-  "scripts": {
-    "start": "npm run gulp && NODE_ENV=production node src/server/index.js",
-    "dev": "npm run gulp && NODE_ENV=development node src/server/index.js",
-    "build": "next build",
-    "gulp": "gulp",
-    "gulp:watch": "gulp watch",
-    "lint": "eslint .",
-    "lint:fix": "eslint . --fix"
-  },
-  "author": "Kumar Abhirup",
-  "repository": {
-    "url": "https://github.com/iqubex-technologies/adify"
-  },
-  "license": "MIT",
-  "dependencies": {
-    "dotenv": "^8.0.0",
-    "express": "^4.17.1",
-    "graphql": "^14.4.2",
-    "graphql-tag": "^2.10.1",
-    "next": "^9.0.3",
-    "npm-run-all": "^4.1.5",
-    "path": "^0.12.7",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0"
-  },
-  "devDependencies": {
-    "autoprefixer": "^9.6.0",
-    "babel-eslint": "^9.0.0",
-    "babel-jest": "^24.8.0",
-    "cssnano": "^4.1.10",
-    "gulp": "^4.0.2",
-    "gulp-postcss": "^8.0.0",
-    "gulp-sass": "^4.0.2",
-    "gulp-sourcemaps": "^2.6.5",
-    "html-webpack-plugin": "^3.2.0",
-    "lerna": "^3.16.4",
-    "node-sass": "^4.12.0",
-    "nodemon": "^1.19.1",
-    "now": "^16.1.1",
-    "prettier": "^1.18.2",
-    "sass": "^1.22.1",
-    "webpack": "^4.35.0",
-    "webpack-cli": "^3.3.5",
-    "webpack-dev-server": "^3.7.2",
-    "eslint": "^5.16.0",
-    "eslint-config-airbnb": "^17.1.1",
-    "eslint-config-prettier": "^4.3.0",
-    "eslint-config-wesbos": "0.0.19",
-    "eslint-plugin-html": "^5.0.5",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prettier": "^3.1.0",
-    "eslint-plugin-react": "^7.14.3",
-    "eslint-plugin-react-hooks": "^1.7.0"
-  }
-}
