--- web/src/__tests__/index.test.js
@@ -1,4 +1,0 @@
-// Slience is golden
-test('should test', () => {
-  expect(1).toBe(1)
-})
