--- web/src/lib/sendFilePath.js
@@ -1,5 +1,0 @@
-const path = require('path')
-
-const sendFilePath = pathName => path.join(__dirname, '../public', pathName)
-
-module.exports = sendFilePath
