--- web/src/public/assets/styles/app.scss
@@ -1,3 +1,0 @@
-body {
-  font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif
-}
