--- web/src/public/index.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <meta charset="utf-8" />
-  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-  <title>Node.js boilerplate</title>
-  <meta name="viewport" content="width=device-width,initial- scale=1" />
-
-  <link rel="stylesheet" href="/public/assets/styles/compiled/app.css" />
-</head>
-
-<body>
-  <h1>Hello World!</h1>
-  <hr />
-  <p>This was boilerplate code setup. Now, it's your chance to make a change!</p>
-</body>
-
-</html>
