--- web/src/server/index.js
@@ -1,27 +1,0 @@
-const express = require('express')
-const next = require('next')
-
-const dev = process.env.NODE_ENV !== 'production'
-const app = next({ dev })
-const handle = app.getRequestHandler()
-
-const PORT = process.env.PORT || 3001
-
-app
-  .prepare()
-  .then(() => {
-    const server = express()
-
-    server.use('/public', express.static('src/public'))
-
-    server.get('*', (req, res) => handle(req, res))
-
-    server.listen(PORT, err => {
-      if (err) throw err
-      console.log(`> Ready on http://localhost:${PORT}`)
-    })
-  })
-  .catch(ex => {
-    console.error(ex.stack)
-    process.exit(1)
-  })
